requirements.txt
elevenlabs
python-dotenv
typer
rich
